JSのファイルの圧縮、ダイレクトアップロード機能を分離
